Fixed Restart Button, changed how audio works for player
The restart panel just needed to function always and not the other thing, then that needed to set the player back when it reset so in global, I set the starting stats again because we may want multiple levels so I didnt want to set it everytime. I also added a 2 simple buttons for the title screen. We have no fonts or button assets so its basic at the moment. Also the audio manager and sound player are weird because it looks like they create themselves everytime?